+++ releasenotes.md
@@ -4,6 +4,7 @@
 ### New Features
 - Add instructions to README.md describing how to build Faker from sources.
 - Class Job supports all 51 built-in as well as custom locales
+- Class Lorem supports all 51 built-in as well as custom locales
 
 ## Version 0.38.0
 ### Breaking Change